Fix pr comments and some cleanup
