+++ internal/systray/available_nonlinux.go
@@ -1,3 +1,5 @@
+//go:build !linux
+
 package systray
 
 func Available() bool { return true }
