+++ internal/systray/systray.go
@@ -13,8 +13,6 @@ import (
 	"runtime"
 	"sync"
 	"sync/atomic"
-
-	"github.com/godbus/dbus/v5"
 )
 
 var (
@@ -118,20 +116,4 @@ func (item *menuItem) update() {
 	menuItems[item.id] = item
 	menuItemsLock.Unlock()
 	addOrUpdateMenuItem(item)
-}
-
-// linux specific check for system tray availability.
-func Available() bool {
-	conn, err := dbus.SessionBus()
-	if err != nil {
-		return false
-	}
-
-	obj := conn.Object("org.kde.StatusNotifierWatcher", "/StatusNotifierWatcher")
-	v, err := obj.GetProperty("org.kde.StatusNotifierWatcher.IsStatusNotifierHostRegistered")
-	if err != nil {
-		return false
-	}
-	registered, ok := v.Value().(bool)
-	return registered && ok
 }
