+++ internal/systray/systray_linux.go
@@ -52,14 +52,6 @@ func (item *menuItem) SetTitle(title string) {
 	item.update()
 }
 
-// SetTooltip sets the systray tooltip to display on mouse hover of the tray icon,
-// only available on Mac and Windows.
-//
-// NOTE: tooltip is not set on Linux.
-func (item *menuItem) SetTooltip(tooltip string) {
-	C.setTooltip(C.CString(tooltip))
-}
-
 func addOrUpdateMenuItem(item *menuItem) {
 	var disabled C.short
 	if item.disabled {
